chore(enums): renamed namespace because of reserved word (insight.symfony)
