+++ tests/Monitoring/Sensor/Service/Mercure.php
@@ -9,7 +9,7 @@ use PHPUnit\Framework\MockObject\MockObject;
 use Symfony\Bundle\MercureBundle\MercureBundle;
 use Symfony\Component\Mercure\Exception\RuntimeException;
 use Symfony\Component\Mercure\HubInterface;
-use whatwedo\MonitorBundle\Enum\SensorStateEnum;
+use whatwedo\MonitorBundle\Enums\SensorStateEnum;
 use whatwedo\MonitorBundle\Monitoring\Sensor\Service\Mercure;
 use whatwedo\MonitorBundle\Tests\Monitoring\AbstractMonitoring;
 use whatwedo\MonitorBundle\Tests\UseTestKernelTrait;
